+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+Under construction.
+
+
+## [0.8.1] - 2023-01-17
+
 ### Added
 
 - Added `Signature::to_be_bytes()`, `Capsule::to_bytes_simple()`, and `CapsuleFrag::to_bytes_simple()` to use in Ethereum contracts. ([#115])
@@ -231,7 +236,7 @@ the corresponding methods in Python and WASM bindings. ([#84])
 
 - Initial release.
 
-[Unreleased]: https://github.com/nucypher/rust-umbral/compare/v0.8.0...HEAD
+[Unreleased]: https://github.com/nucypher/rust-umbral/compare/v0.8.1...HEAD
 [0.2.0]: https://github.com/nucypher/rust-umbral/releases/tag/v0.2.0
 [0.3.0]: https://github.com/nucypher/rust-umbral/releases/tag/v0.3.0
 [0.4.0]: https://github.com/nucypher/rust-umbral/releases/tag/v0.4.0
@@ -241,3 +246,4 @@ the corresponding methods in Python and WASM bindings. ([#84])
 [0.6.0]: https://github.com/nucypher/rust-umbral/releases/tag/v0.6.0
 [0.7.0]: https://github.com/nucypher/rust-umbral/releases/tag/v0.7.0
 [0.8.0]: https://github.com/nucypher/rust-umbral/releases/tag/v0.8.0
+[0.8.1]: https://github.com/nucypher/rust-umbral/releases/tag/v0.8.1